Fix some warnings (#2006)
* Remove extra brackets

* Simplify a case statement in quota code.

Condition is always 'true' because 'prev' is always not 'nil'

If prev was nill the previous case would have handled it.

* Avoid potential (unlikely) nil pointer dereference in quota test

* Remove a couple of redundant type conversions

* Fix quota test error handling + convert to t.Run()

* Make the error messages more standard

* And another message

* Update quota/etcd/storage/quota_storage_test.go


* Slight simplification as the test only handles error cases.

* Make tests in t.Run() independent, remove comment that they're not

+++ merkle/compact/range_test.go
@@ -588,7 +588,7 @@ func TestDecomposeCases(t *testing.T) {
 func verifyDecompose(begin, end uint64) error {
 	left, right := Decompose(begin, end)
 	// Smoke test the sum of decomposition masks.
-	if left+right != uint64(end-begin) {
+	if left+right != end-begin {
 		return fmt.Errorf("%d+%d != %d-%d", left, right, begin, end)
 	}
